Add SQLite support for testing purposes

+++ prepare_db.py
@@ -30,6 +30,11 @@ APPS = (
 
 if __name__ == '__main__':
     db_info = DATABASES['default']
+    if 'sqlite' in db_info['ENGINE']:
+        print('Re-creating SQLite database file "{}"... '
+              ''.format(db_info['NAME']), end='')
+        open(db_info['NAME'], mode='w').close()
+        print('OK')
     if 'postgresql' in db_info['ENGINE']:
         import psycopg2
         connection = psycopg2.connect(
@@ -42,13 +47,14 @@ if __name__ == '__main__':
         os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'schoolmate.settings')
         from django.db import connection
         cursor = connection.cursor()
-    print('Deleting database "{}"... '.format(db_info['NAME']), end='')
-    cursor.execute('DROP DATABASE {}'.format(db_info['NAME']))
-    print('OK')
-    print('Re-creating database... ', end='')
-    cursor.execute('CREATE DATABASE {} WITH ENCODING \'UTF8\''
-                   ''.format(db_info['NAME']))
-    print('OK')
+    if 'sqlite' not in db_info['ENGINE']:
+        print('Deleting database "{}"... '.format(db_info['NAME']), end='')
+        cursor.execute('DROP DATABASE {}'.format(db_info['NAME']))
+        print('OK')
+        print('Re-creating database... ', end='')
+        cursor.execute('CREATE DATABASE {} WITH ENCODING \'UTF8\''
+                       ''.format(db_info['NAME']))
+        print('OK')
     print('Re-creating migration directories... ', end='')
     dirs = [os.path.join(BASE_DIR, app, 'migrations') for app in APPS]
     [shutil.rmtree(d) if os.path.exists(d) else None for d in dirs]
